Byte collection cleanups
