+++ academy-ui/src/app/clickable-demos/view/[...demoId]/page.tsx
@@ -2,7 +2,8 @@
 
 import ClickableDemoModal from '@/components/clickableDemos/View/ClickableDemoModal';
 import withSpace from '@/contexts/withSpace';
-import { ClickableDemoWithSteps, SpaceWithIntegrationsFragment, SpaceTypes } from '@/graphql/generated/generated-types';
+import { ClickableDemoWithSteps, SpaceWithIntegrationsFragment } from '@/graphql/generated/generated-types';
+import { SpaceTypes } from '@/types/space/SpaceDto';
 import FullPageLoader from '@dodao/web-core/components/core/loaders/FullPageLoading';
 import getBaseUrl from '@dodao/web-core/utils/api/getBaseURL';
 import axios from 'axios';
@@ -42,7 +43,7 @@ function ViewClickableDemo({ params, space }: { params: { demoId: string[] }; sp
         clickableDemoWithSteps={data}
         space={space}
         onClose={() => {
-          router.push(`${space.type === SpaceTypes.TidbitsSite ? '/' : '/tidbit-collections'}`);
+          router.push(`${space.type === SpaceTypes.TidbitsSite ? '/' : '/clickable-demos'}`);
         }}
       />
     );
