+++ academy-ui/src/components/byteCollection/ByteCollections/AddByteCollection.tsx
@@ -1,18 +1,11 @@
 'use client';
-import ByteCollectionEditor from '@/components/byteCollection/ByteCollections/ByteCollectionEditor';
+import ByteCollectionEditModal from '@/components/byteCollection/ByteCollections/ByteCollectionEditModal';
 import { SpaceWithIntegrationsFragment } from '@/graphql/generated/generated-types';
-import SingleCardLayout from '@/layouts/SingleCardLayout';
 import Button from '@dodao/web-core/components/core/buttons/Button';
-import FullScreenModal from '@dodao/web-core/components/core/modals/FullScreenModal';
-import PageWrapper from '@dodao/web-core/components/core/page/PageWrapper';
 import React from 'react';
 
 export default function AddByteCollection({ space }: { space: SpaceWithIntegrationsFragment }) {
   const [showAddCollectionModal, setShowAddCollectionModal] = React.useState(false);
-
-  function onClose() {
-    setShowAddCollectionModal(false);
-  }
 
   return (
     <div className="w-full flex justify-center">
@@ -26,15 +19,13 @@ export default function AddByteCollection({ space }: { space: SpaceWithIntegrati
         + Add Tidbit Collection
       </Button>
       {showAddCollectionModal && (
-        <FullScreenModal open={true} onClose={onClose} title={'Create Tidbit Collection'}>
-          <div className="text-left">
-            <PageWrapper>
-              <SingleCardLayout>
-                <ByteCollectionEditor space={space} byteCollection={undefined} viewByteCollectionsUrl={'/'} />
-              </SingleCardLayout>
-            </PageWrapper>
-          </div>
-        </FullScreenModal>
+        <ByteCollectionEditModal
+          space={space}
+          byteCollection={undefined}
+          onClose={() => {
+            setShowAddCollectionModal(false);
+          }}
+        />
       )}
     </div>
   );
