+++ academy-ui/src/types/space/SpaceDto.ts
@@ -1,0 +1,4 @@
+export enum SpaceTypes {
+  AcademySite = 'ACADEMY_SITE',
+  TidbitsSite = 'TidbitsSite',
+}
